pyproject.toml
# SPDX-FileCopyrightText: 2026 Jose Gonzalo Covarrubias
[project]
name = "xx2html"
authors = [
    { name = "Jose Gonzalo Covarrubias", email = "gocova.dev+xx2html@gmail.com" },
]
description = "Convert XLSX workbooks to single styled HTML file with support for conditional formatting and in-cell images."
license = "MIT OR Apache-2.0"
license-files = ["LICEN[CS]E*"]
readme = "README.md"
dependencies = [
    "bs4>=0.0.2",
    "condif2css>=0.13.0b6021601",
    "lxml>=5.3.1",
    "mvin>=0.13.0b6021601",
    "openpyxl>=3.1.5",
    "pillow>=11.0.0",
    "xlsx2html>=0.6.1",
]
dynamic = ["version"]
requires-python = ">=3.10,<4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Development Status :: 4 - Beta",
]

[project.urls]
Homepage = "https://bitbucket.org/gocova_dev/xx2html"
Repository = "https://bitbucket.org/gocova_dev/xx2html"
Issues = "https://bitbucket.org/gocova_dev/xx2html/issues"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
version = { source = "file", path = "src/xx2html/_constants.py" }
distribution = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "types-lxml>=2026.2.16",
    "types-openpyxl>=3.1.5.20250919",
]

[tool.coverage.run]
branch = true
source = ["src/xx2html"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=xx2html --cov-report=term-missing --cov-fail-under=60"
